+++ backend/forgather/src/main/java/com/forgather/domain/upload/dto/IssueSignedUrlRequest.java
@@ -1,4 +1,4 @@
-package com.forgather.domain.space.dto;
+package com.forgather.domain.upload.dto;
 
 import java.util.List;
 
@@ -8,7 +8,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public record IssueSignedUrlRequest(
 
-    @Schema(description = "업로드 파일 카테고리", example = "\"profile\"")
+    @Schema(description = "업로드 파일 카테고리", examples = {"\"PRODUCT\"", "\"GUESTBOOK\""})
     UploadCategory category,
 
     @Schema(description = "업로드 파일 이름 목록", example = "[\"UUID1.png\", \"UUID2.jpeg\", \"UUID3.png\"]")
